+++ tests/python_tests/render_grid_test.py
@@ -57,6 +57,8 @@ def create_grid_map(width,height):
     s = mapnik.Style()
     r = mapnik.Rule()
     symb = mapnik.MarkersSymbolizer()
+    symb.width = 10
+    symb.height = 10
     symb.allow_overlap = True
     r.symbols.append(symb)